+++ randomizer-data/src/test/java/de/alpharogroup/random/SecureRandomBeanTest.java
@@ -24,11 +24,12 @@
  */
 package de.alpharogroup.random;
 
+import static org.testng.AssertJUnit.assertNotNull;
+
 import java.security.NoSuchAlgorithmException;
 import java.security.NoSuchProviderException;
 import java.security.SecureRandom;
 
-import org.testng.AssertJUnit;
 import org.testng.annotations.Test;
 
 /**
@@ -53,33 +54,10 @@ public class SecureRandomBeanTest
 		SecureRandom sr = SecureRandomBean.builder().build();
 
 		sr = SecureRandomBean.builder().algorithm(SecureRandomBean.DEFAULT_ALGORITHM).build();
-		AssertJUnit.assertNotNull(sr);
+		assertNotNull(sr);
 		sr = SecureRandomBean.builder().algorithm(SecureRandomBean.DEFAULT_ALGORITHM)
 			.provider("SUN").build();
-		AssertJUnit.assertNotNull(sr);
-	}
-
-	/**
-	 * Test method for {@link SecureRandomBean#buildQuietly()}
-	 */
-	@Test
-	public void testBuildQuietly()
-	{
-		SecureRandom sr;
-		sr = SecureRandomBean.builder().buildQuietly();
-
-		sr = SecureRandomBean.builder().algorithm(SecureRandomBean.DEFAULT_ALGORITHM)
-			.buildQuietly();
-		AssertJUnit.assertNotNull(sr);
-		sr = SecureRandomBean.builder().algorithm(SecureRandomBean.DEFAULT_ALGORITHM)
-			.provider("SUN").buildQuietly();
-		AssertJUnit.assertNotNull(sr);
-
-		sr = SecureRandomBean.builder().algorithm("FOO").buildQuietly();
-		AssertJUnit.assertNull(sr);
-		sr = SecureRandomBean.builder().algorithm(SecureRandomBean.DEFAULT_ALGORITHM)
-			.provider("BAR").buildQuietly();
-		AssertJUnit.assertNull(sr);
+		assertNotNull(sr);
 	}
 
 	/**